Group TestUtil functions into separate classes
