+++ src/test/java/guitests/guihandles/PersonListPanelHandle.java
@@ -15,7 +15,8 @@ import javafx.stage.Stage;
 import seedu.address.TestApp;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.ReadOnlyPerson;
-import seedu.address.testutil.TestUtil;
+import seedu.address.testutil.TestPersonCardUtil;
+import seedu.address.testutil.TestNodeUtil;
 
 /**
  * Provides a handle for the panel containing the person list.
@@ -63,7 +64,7 @@ public class PersonListPanelHandle extends GuiHandle {
             final int scrollTo = i + startPosition;
             guiRobot.interact(() -> getListView().scrollTo(scrollTo));
             guiRobot.sleep(200);
-            if (!TestUtil.isPersonSameAsPersonOnCard(getPersonCardHandle(startPosition + i), persons[i])) {
+            if (!TestPersonCardUtil.isPersonSameAsPersonOnCard(getPersonCardHandle(startPosition + i), persons[i])) {
                 return false;
             }
         }
@@ -74,7 +75,7 @@ public class PersonListPanelHandle extends GuiHandle {
      * Clicks on the ListView.
      */
     public void clickOnListView() {
-        Point2D point = TestUtil.getScreenMidPoint(getListView());
+        Point2D point = TestNodeUtil.getScreenMidPoint(getListView());
         guiRobot.clickOn(point.getX(), point.getY());
     }
