+++ src/test/java/seedu/address/commons/util/XmlUtilTest.java
@@ -14,7 +14,8 @@ import org.junit.rules.ExpectedException;
 import seedu.address.model.AddressBook;
 import seedu.address.storage.XmlSerializableAddressBook;
 import seedu.address.testutil.AddressBookBuilder;
-import seedu.address.testutil.TestUtil;
+import seedu.address.testutil.TemporaryFilesUtil;
+import seedu.address.testutil.MockPersonsData;
 
 public class XmlUtilTest {
 
@@ -22,7 +23,7 @@ public class XmlUtilTest {
     private static final File EMPTY_FILE = new File(TEST_DATA_FOLDER + "empty.xml");
     private static final File MISSING_FILE = new File(TEST_DATA_FOLDER + "missing.xml");
     private static final File VALID_FILE = new File(TEST_DATA_FOLDER + "validAddressBook.xml");
-    private static final File TEMP_FILE = new File(TestUtil.getFilePathInSandboxFolder("tempAddressBook.xml"));
+    private static final File TEMP_FILE = new File(TemporaryFilesUtil.getFilePathInSandboxFolder("tempAddressBook.xml"));
 
     @Rule
     public ExpectedException thrown = ExpectedException.none();
@@ -87,7 +88,7 @@ public class XmlUtilTest {
 
         AddressBookBuilder builder = new AddressBookBuilder(new AddressBook());
         dataToWrite = new XmlSerializableAddressBook(
-                builder.withPerson(TestUtil.getSamplePersonsList().get(0)).withTag("Friends").build());
+                builder.withPerson(MockPersonsData.getSamplePersonsList().get(0)).withTag("Friends").build());
 
         XmlUtil.saveDataToFile(TEMP_FILE, dataToWrite);
         dataFromFile = XmlUtil.getDataFromFile(TEMP_FILE, XmlSerializableAddressBook.class);
