+++ src/test/java/seedu/address/testutil/TestKeyCodeUtil.java
@@ -1,0 +1,35 @@
+package seedu.address.testutil;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import javafx.scene.input.KeyCode;
+import javafx.scene.input.KeyCodeCombination;
+import javafx.scene.input.KeyCombination;
+
+/**
+* Utility class for key code testing.
+*/
+public class TestKeyCodeUtil {
+    /**
+     * Tweaks the {@code keyCodeCombination} to resolve the {@code KeyCode.SHORTCUT} to their
+     * respective platform-specific keycodes
+     */
+    public static KeyCode[] scrub(KeyCodeCombination keyCodeCombination) {
+        List<KeyCode> keys = new ArrayList<>();
+        if (keyCodeCombination.getAlt() == KeyCombination.ModifierValue.DOWN) {
+            keys.add(KeyCode.ALT);
+        }
+        if (keyCodeCombination.getShift() == KeyCombination.ModifierValue.DOWN) {
+            keys.add(KeyCode.SHIFT);
+        }
+        if (keyCodeCombination.getMeta() == KeyCombination.ModifierValue.DOWN) {
+            keys.add(KeyCode.META);
+        }
+        if (keyCodeCombination.getControl() == KeyCombination.ModifierValue.DOWN) {
+            keys.add(KeyCode.CONTROL);
+        }
+        keys.add(keyCodeCombination.getCode());
+        return keys.toArray(new KeyCode[]{});
+    }
+}
